Refactored data fetching within individual components and Implemented Suspense
